+++ index.ts
@@ -1,2 +1,2 @@
 export * from './credentials/AzuraCastApi.credentials';
-export * from './nodes/AzuraCast/AzuraCast.node';
+export * from './nodes/AzuraCast';